// web-backup/lib/types/football.ts
export interface ResponseParameters {
  league?: string
  season?: string
  date?: string
  from?: string
  to?: string
  team?: string
  fixture?: string
  live?: string
  next?: string
  last?: string
}

export interface APIPaging {
  current: number
  total: number
}

// Fixture Types
export interface FixturesResponse {
  get: string
  parameters: ResponseParameters
  errors: any
  results: number
  paging: APIPaging
  response: Fixture[]
}

export interface Fixture {
  fixture: FixtureDetails
  league: LeagueFixtureInfo
  teams: Teams
  goals: Goals | null
}

export interface FixtureDetails {
  id: number
  date: string
  status: FixtureStatus
  venue: Venue
  timezone: string
  referee: string | null
}

export interface FixtureStatus {
  long: string
  short: string
  elapsed: number | null
}

export interface Venue {
  id: number | null
  name: string | null
  city: string | null
}

export interface LeagueFixtureInfo {
  id: number
  name: string
  country: string
  logo: string
  flag: string | null
  season: number
  round: string
  standings: boolean | null
}

export interface Teams {
  home: Team
  away: Team
}

export interface Team {
  id: number
  name: string
  logo: string
  winner: boolean | null
}

export interface Goals {
  home: number | null
  away: number | null
}